refactor: Add save_ice_mask_hist function for saving ice mask histogram
Add a new function `save_ice_mask_hist` to the `savefigs.py` file. This function generates a histogram of the `red_masked` data, with specified bins, minimum and maximum cutoff values, and color. The histogram is saved as an image file in the target directory with the name `{doy}_ice_mask_hist.png`. This function is useful for visualizing the distribution of ice mask values.